fixed and working usb spreader, now also mutates when ran from the different expected malware path
