--- src/main/modules/usb_spread_module/string_utils.h
@@ -1,8 +1,0 @@
-#pragma once
-
-char *findSubstring(char* string, char *substr);
-
-char *concat(const char *s1, const char *s2);
-
-// To fix "global" variable, issue with loading crashes
-void init_string_utils(Globals * glob);
